requirements.txt
flask
pyjwt
werkzeug
flask-cors
python-dotenv